DatLA add new source code for function 5.4.1 show list of notification and delete notification date 03/02/2023.

+++ C0722G1-Repo-BE/src/main/java/com/c0722g1repobe/controller/notification/NotificationRestController.java
@@ -43,6 +43,7 @@ public class NotificationRestController {
         return new ResponseEntity<>(notificationPage, HttpStatus.OK);
     }
 
+
     /**
      * Create by: DatLA
      * Date created: 31/01/2023
@@ -52,7 +53,7 @@ public class NotificationRestController {
      * @return HttpStatus.NO_CONTENT if exists any notification not found or HttpStatus.OK and notification found
      */
     @PostMapping("/find-by-list-id")
-    public ResponseEntity<List<NotificationDeleteDto>> findByListId(@RequestBody List<Integer> idList) {
+    public ResponseEntity<List<NotificationDeleteDto>> findByListId(@RequestBody List<Long> idList) {
         if (idList == null || idList.isEmpty()) {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
@@ -73,7 +74,7 @@ public class NotificationRestController {
      * or HttpStatus.BAD_REQUEST if request is error
      */
     @PostMapping("/remove")
-    public ResponseEntity<HttpStatus> remove(@RequestBody List<Integer> idList) {
+    public ResponseEntity<HttpStatus> remove(@RequestBody List<Long> idList) {
         if (idList == null || idList.isEmpty()) {
             return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
         }
